parser: fix: save a raw message slice in PSIPMsg
PSIPMsg used to contain only a slice pointing to the actual raw
message. This slice (Buf) is used to retrieve header values and
other strings from the message (each of the values are stored as
an (offset, len) pair inside the original data passed for
parsing).
However if the initial offset passed to ParseSIPMsg() was not 0,
attempting to retrieve some value would produce wrong results,
since Buf[] is adjusted to point to the actual message start, but
all the header values are relative to the data start.

Fix: make Buf[] always point at the beginning of the data passed
for parsing and have a different RawMsg[] slice that always points
to the actual raw message.
Minimal impact, since so far all the code uses 0 offset in the
initial parse call.