fix: parse split body returned wrong offset
If the provided message buffer does not contain the whole body the
parsing function will return ErrHdrMoreBytes and a continuation
offset (supposed to be used to continue parsing after more data is
added to the buffer). However in the body case, the offset
returned was always the end of the buffer which was wrong because
there is no internal body-offset state kept and the end of the
body is found using the content-length + passed offset => a
message with a partial body will cause the offset to get
out-of-sync (particularly bad on streams like tcp).
The fix works by always returning the start-of-body offset if the
body is not fully contained in the message buffer (an alternative
fix would be to keep the start of body offset in the internal state
 and use it instead of the passed "current" offset for the end of
 body calculations).