message signature support
Support for generating message signatures for sip requests.
For now the signature contains the method and the order of some
common headers.
The ascii format is: method hdr_sig_id... , represented as a hex
string (the method and each head hdr_sig_id are a all represented
as a single hex digit).
The hdr_sig_id is a 4-bit number consisting of  the header name
 size (1 for compact, 0 for normal) and a 3-bit header type.
The headers used for the header order signature are:
Call-ID, Contact, CSeq, From, Max-Forwads, To, Via (1st) and
User-Agent.

Example: 26352041  for INVITE with the following header order:
 Via, From, To, CSeq, Call-ID, Max-Forwards, Contact.
If the Via, From, To and Call-ID are in compact form (v, f, t, i),
then the signature would be: 2ebd2841.